Add call to backend for books, completed add to readlist functionality

+++ project2-angular/src/app/components/readlist/readlist.component.ts
@@ -3,7 +3,6 @@ import { ReadlistService } from 'src/app/services/readlist.service';
 import { IReadlist } from 'src/app/models/readlist.model';
 import { UserService } from 'src/app/user/user.service';
 import { User } from 'src/app/user/user';
-import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-readlist',
@@ -14,10 +13,9 @@ export class ReadlistComponent implements OnInit {
   readlists: IReadlist[] = [];
   userLoggedIn: boolean = false;
   userHasReadlist: boolean = false;
-  userId: number = 0;
   currentUser: User = {
     userId: 0,
-    firstName : "test",
+    firstName: "test",
     lastName: "user",
     mobile: "",
     email: "",
@@ -25,29 +23,34 @@ export class ReadlistComponent implements OnInit {
     password: ""
   };
 
-
-  constructor(private readlistService: ReadlistService, private userService: UserService) { }
+  constructor(private readlistService: ReadlistService,
+    private userService: UserService) { }
 
   ngOnInit(): void {
     this.userLoggedIn = false;
     this.userHasReadlist = false;
-    this.userId = this.userService.getCurrentUser().userId;
     this.currentUser = this.userService.getCurrentUser();
 
-    if (this.userId != 0) {
+    if (this.currentUser.userId != 0) {
       this.userLoggedIn = true;
     }
-    
-    this.readlistService.getReadlistById(this.userId).subscribe(data => {
+
+    this.readlistService.getReadlistById(this.currentUser.userId).subscribe(data => {
       this.readlists = data;
       if (data.length != 0) {
         this.userHasReadlist = true;
       }
-    })
+    });
   }
 
   public removeBook(data: IReadlist) {
-    this.readlistService.deleteReadlistEntry(data.readListId)
-      .subscribe(() => location.reload());
+    if (data.readListId != undefined) {
+      this.readlistService.deleteReadlistEntry(data.readListId)
+        .subscribe(() => this.ngOnInit());
+    }
+  }
+
+  public addBook(bookISBN: number) { 
+    this.readlistService.addReadlistEntry(bookISBN);
   }
 }
